Init commit, refactor dep to shareable component
