+++ service/frontend/nexus_http_handler.go
@@ -56,6 +56,7 @@ func NewNexusHTTPHandler(
 	endpointRegistry commonnexus.EndpointRegistry,
 	authInterceptor *authorization.Interceptor,
 	telemetryInterceptor *interceptor.TelemetryInterceptor,
+	errorHandler *interceptor.RequestErrorHandler,
 	redirectionInterceptor *interceptor.Redirection,
 	namespaceValidationInterceptor *interceptor.NamespaceValidatorInterceptor,
 	namespaceRateLimitInterceptor interceptor.NamespaceRateLimitInterceptor,
@@ -84,6 +85,7 @@ func NewNexusHTTPHandler(
 				matchingClient:                matchingClient,
 				auth:                          authInterceptor,
 				telemetryInterceptor:          telemetryInterceptor,
+				errorHandler:                  errorHandler,
 				redirectionInterceptor:        redirectionInterceptor,
 				forwardingEnabledForNamespace: serviceConfig.EnableNamespaceNotActiveAutoForwarding,
 				forwardingClients:             clientCache,
